app/requirements.txt
asgiref==3.4.1
Django==3.2.5
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
Unidecode==1.2.0
celery==5.1.2
celery-once==3.0.1
dnspython==2.1.0
psycopg2-binary==2.9.1
redis==3.5.3
selenium==3.141.0
gunicorn
djangorestframework
django_celery_beat
django-cors-headers
django-mathfilters

